langgraph-ai/pyproject.toml
[project]
name = "contract-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=1.2.17",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "openai>=2.24.0",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pypdf>=6.7.5",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
    "tiktoken>=0.12.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
